OAK-11930 improve the output, add testcases
